Renamed source/detail/time.hpp to time_helpers.hpp based on feedback.
